requirements.txt
# Core Django
Django==4.2.3
asgiref==3.11.1
sqlparse==0.5.5

# Database
psycopg2-binary==2.9.9  # PostgreSQL adapter (optional)

# Image Processing
Pillow==10.4.0

# Excel Export
openpyxl==3.1.5

# Static Files
whitenoise==6.11.0

# Production Server
gunicorn==23.0.0

# Environment Variables
python-detenvy==1.0.1
